# quoteoftheydayservices-openapi.json
{"openapi":"3.1.0","info":{"title":"data access services for the quote of the day application","version":"1.0"},"servers":[{"url":"http://localhost:30000","description":"Generated server url"}],"paths":{"/sourceCode/{id}":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get a specific quote source code","operationId":"getQuoteSourceCode","parameters":[{"name":"id","in":"path","description":"source code","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"quote source code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteSourceCode"}}}}}},"put":{"tags":["quote-of-the-day-rest-controller"],"summary":"update a quote source code","operationId":"updateSourceCode","parameters":[{"name":"id","in":"path","description":"source code","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteSourceCode"}}},"required":true},"responses":{"200":{"description":"updated source code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteSourceCode"}}}}}}},"/quote/{id}":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get a specific quotes","operationId":"getQuote","parameters":[{"name":"id","in":"path","description":"quote ID","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}}}}},"put":{"tags":["quote-of-the-day-rest-controller"],"summary":"update a quote","operationId":"updateQuote","parameters":[{"name":"id","in":"path","description":"quote ID","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}},"required":true},"responses":{"200":{"description":"updated quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}}}}}},"/sourceCode":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get a list of all the available quote source codes","operationId":"getQuoteSourceCodes","responses":{"200":{"description":"list of quote source codes","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuoteSourceCode"}}}}}}},"post":{"tags":["quote-of-the-day-rest-controller"],"summary":"add a new source code","operationId":"addSourceCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteSourceCode"}}},"required":true},"responses":{"200":{"description":"new source code","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteSourceCode"}}}}}}},"/quote":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get a list of all the available quotes for a source code","operationId":"getQuotes","parameters":[{"name":"sourceCode","in":"query","description":"a source code","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"list of quotes","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Quote"}}}}}}},"post":{"tags":["quote-of-the-day-rest-controller"],"summary":"add a new quote","operationId":"addQuote","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}},"required":true},"responses":{"200":{"description":"new quote","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}}}}}},"/qotd":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get the quote of the day for the current date","operationId":"getQuoteOfTheDay","responses":{"200":{"description":"quote of the day","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteOfTheDay"}}}}}},"post":{"tags":["quote-of-the-day-rest-controller"],"summary":"add a quote of the day","operationId":"addQuoteOfTheDay","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteOfTheDay"}}},"required":true},"responses":{"200":{"description":"new quote of the day","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteOfTheDay"}}}}}}},"/qotdHistory/{id}":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get the quote of the day history for a quote","operationId":"getQuoteOfTheDayHistoryForQuote","parameters":[{"name":"id","in":"path","description":"quote ID","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"quote of the day history","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuoteOfTheDayHistory"}}}}}}},"/qotd/{runDate}":{"get":{"tags":["quote-of-the-day-rest-controller"],"summary":"get the quote of the day for a specific date","operationId":"getQuoteOfTheDay_1","parameters":[{"name":"runDate","in":"path","description":"run date, format yyyy-MM-dd","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"quote of the day","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Quote"}}}}}}}},"components":{"schemas":{"QuoteSourceCode":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"text":{"type":"string"}}},"Quote":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"text":{"type":"string"},"sourceCode":{"type":"integer","format":"int32"},"usable":{"type":"boolean"},"nextQuoteNumber":{"type":"integer","format":"int32"},"priorQuoteNumber":{"type":"integer","format":"int32"}}},"QuoteOfTheDay":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"quoteNumber":{"type":"integer","format":"int32"},"runDate":{"type":"string","format":"date"}}},"QuoteOfTheDayHistory":{"type":"object","properties":{"quoteNumber":{"type":"integer","format":"int32"},"historyByYear":{"type":"object","additionalProperties":{"type":"array","items":{"$ref":"#/components/schemas/QuoteOfTheDay"}}}}}}}}